Moved `SharedType`, `SharedReference`, and `SharedObject` to `shared_object.dart`. ; Renamed `isIsolateCopy` to `isAuxiliaryCopy`.
